+++ Test/TestIMovieRatingRepository.cs
@@ -59,7 +59,7 @@ namespace Test
         public void TestGetAverageRatingByMovieName(string movieName, int[] ratings)
         {
             Array.ForEach(ratings, rating => repository.SaveRating(movieName, rating));
-            var expected = Math.Round((decimal)ratings.Average(), 3);
+            var expected = Math.Round(ratings.Average(), 3);
 
             var result = Math.Round(repository.GetAverageRatingByMovieName(movieName), 3);
